requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.5
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
frozenlist==1.4.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kombu==5.3.2
MarkupSafe==2.1.3
multidict==6.0.4
openapi==1.1.0
openapi-codec==1.3.2
packaging==23.1
prompt-toolkit==3.0.39
psycopg==3.1.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
